+++ src/components/Dashboard/Header.jsx
@@ -10,11 +10,11 @@ const Header = ({ role }) => {
   const [logout, setLogout] = useState(false);
   const toggleLogout = () => setLogout(!logout);
   return (
-    <header className="h-20 w-full border-b border-[#dddddd] flex justify-between items-center bg-[#F7F9FC] px-5 sm:px-5 xl:px-10">
+    <header className="h-15 sm:h-20 w-full border-b border-[#dddddd] flex justify-between items-center bg-[#F7F9FC] px-5 sm:px-5 xl:px-10">
       <div className="flex items-center lg:w-60">
-        <span className="text-[20px] font-medium text-[#344054]">
+        <span className="text-[14px] sm:text-[20px] font-medium text-[#344054]">
           {role === "Farmer"
-            ? role + "Dashboard"
+            ? role + " Dashboard"
             : role === "Facility Owner"
             ? role + " Dashboard"
             : "Error 404"}
@@ -30,22 +30,23 @@ const Header = ({ role }) => {
       </div>
       <div className="flex items-center space-y-4 sm:space-y-0 sm:space-x-4 justify-center sm:justify-end">
         <div className="flex items-center h-10 gap-4 p-2 rounded-lg">
-          <div className="h-10 w-10 rounded-full bg-[#D5F0E8] flex justify-center items-center text-xl text-[#047D58]">
+          <Link className="fixed sm:static right-16 h-6 w-6 sm:h-10 sm:w-10 rounded-full bg-[#D5F0E8] flex justify-center items-center sm:text-xl text-[#047D58]">
             <IoNotifications />
-          </div>
+            <span className="text-red-600 relative bottom-[6px] sm:bottom-[8px] right-[3px] text-[9px] sm:text-[12px] font-bold">3</span>
+          </Link>
           <div
             onClick={() => toggleLogout()}
-            className="flex items-center gap-1"
+            className="fixed sm:static top-4 right-4 sm:flex items-center gap-1"
           >
-            <div className="flex items-center space-x-2">
-              <img src={assets.intern} /> <p>Interns</p>
+            <div className="flex items-center space-x-1 sm:space-x-2">
+              <img className="size-8 sm:size-10" src={assets.intern} /> <p className="hidden sm:block">Interns</p>
             </div>
-            <div>
+            <div className="hidden sm:block">
               <FaAngleDown />
             </div>
             <div
               className={`${
-                logout ? "relative" : "hidden"
+                logout ? "opacity-0 sm:relative sm:opacity-100" : "hidden"
               } top-20 right-5 w-0 h-15 rounded-md flex items-center px-2 text-red-600`}
             >
               <Link
