+++ src/components/Dashboard/RevenueData.txt
@@ -1,0 +1,114 @@
+{/* Revenue Chart & Sales Report */}
+                {/* <div className="grid grid-cols-1 lg:grid-cols-2 gap-6 mb-8">
+                <div className="bg-white p-6 rounded-xl shadow-md">
+                  <h2 className="text-xl font-semibold text-gray-900 mb-4">
+                    Revenue
+                  </h2>
+                  <div className="flex items-center space-x-4 mb-4">
+                    <span className="text-3xl font-bold text-gray-900">
+                      ₦87.500
+                    </span>
+                    <div className="flex items-center text-green-600 text-sm font-semibold">
+                      <span className="mr-1">Up 42 %</span>
+                      <span className="text-gray-500">Month-over-month</span>
+                    </div>
+                  </div>
+                  <div className="h-64 w-full">
+                    
+                    <ResponsiveContainer width="100%" height="100%">
+                      <LineChart
+                        data={REVENUE_DATA}
+                        margin={{ top: 5, right: 0, left: -20, bottom: 5 }} // Adjust margins to fit
+                      >
+                        <XAxis
+                          dataKey="name"
+                          tickLine={false}
+                          axisLine={false}
+                          tick={{ fontSize: 12, fill: "#6B7280" }}
+                        />
+                        <YAxis
+                          tickLine={false}
+                          axisLine={false}
+                          tick={{ fontSize: 12, fill: "#6B7280" }}
+                          tickFormatter={(value) =>
+                            `₦${value.toLocaleString()}`
+                          }
+                        />
+                        <Tooltip content={<CustomLineChartTooltip />} />
+                        <Line
+                          type="monotone"
+                          dataKey="value"
+                          stroke="#04D957"
+                          strokeWidth={3}
+                          dot={false}
+                        />
+                      </LineChart>
+                    </ResponsiveContainer>
+                  </div>
+                  <div className="flex justify-around text-xs text-gray-600 mt-4">
+                  </div>
+                </div>
+
+                //Sales Report
+                <div className="bg-white p-6 rounded-xl shadow-md">
+                  <h2 className="text-xl font-semibold text-gray-900 mb-4">
+                    Sales report
+                  </h2>
+                  <div className="space-y-4 mb-6">
+                    //Produce Processed
+                    <div>
+                      <div className="flex justify-between text-sm text-gray-700 mb-1">
+                        <span>Produce Processed</span>
+                        <span>(124)</span>
+                      </div>
+                      <div className="w-full bg-gray-200 rounded-full h-2.5">
+                        <div
+                          className="bg-green-500 h-2.5 rounded-full"
+                          style={{ width: `${(124 / 400) * 100}%` }}
+                        ></div>{" "}
+                        //Scaled width
+                      </div>
+                    </div>
+                    //Ongoing Project
+                    <div>
+                      <div className="flex justify-between text-sm text-gray-700 mb-1">
+                        <span>Ongoing Project</span>
+                        <span>(32)</span>
+                      </div>
+                      <div className="w-full bg-gray-200 rounded-full h-2.5">
+                        <div
+                          className="bg-blue-500 h-2.5 rounded-full"
+                          style={{ width: `${(32 / 400) * 100}%` }}
+                        ></div>
+                      </div>
+                    </div>
+                    //Produce Sold
+                    <div>
+                      <div className="flex justify-between text-sm text-gray-700 mb-1">
+                        <span>Produce Sold</span>
+                        <span>(315)</span>
+                      </div>
+                      <div className="w-full bg-gray-200 rounded-full h-2.5">
+                        <div
+                          className="bg-purple-500 h-2.5 rounded-full"
+                          style={{ width: `${(315 / 400) * 100}%` }}
+                        ></div>
+                      </div>
+                    </div>
+                    <div className="text-xs text-gray-600 mt-4">
+                      <span className="font-semibold mr-1">Net sales</span>
+                      <span className="text-green-600">
+                        Up 42 % Month-over-manth
+                      </span>
+                    </div>
+                  </div>
+                  //Horizontal scale 
+                  <div className="flex justify-between text-xs text-gray-500 mt-6 pt-2 border-t border-gray-200">
+                    <span>0</span>
+                    <span>100</span>
+                    <span>200</span>
+                    <span>300</span>
+                    <span>400</span>
+                  </div>
+                </div>
+              </div> */}
